+++ common/src/main/java/de/maxhenkel/betterrespawn/RespawnAbilities.java
@@ -24,4 +24,6 @@ public interface RespawnAbilities {
 
     boolean getRespawnForced();
 
+    long getLastDeathTime();
+
 }