+++ fabric/src/main/java/de/maxhenkel/betterrespawn/config/FabricServerConfig.java
@@ -1,5 +1,6 @@
 package de.maxhenkel.betterrespawn.config;
 
+
 import de.maxhenkel.configbuilder.ConfigBuilder;
 
 public class FabricServerConfig extends ServerConfig {
@@ -7,24 +8,24 @@ public class FabricServerConfig extends ServerConfig {
     public FabricServerConfig(ConfigBuilder builder) {
         maxRespawnDistance = builder.integerEntry(
                 "max_respawn_distance",
-                256,
+                2500,
                 16,
                 8192,
-                "The maximum distance to spawn the player away from its death location"
+                "Maximum distance for player respawn from the center of the world"
         );
         minRespawnDistance = builder.integerEntry(
                 "min_respawn_distance",
-                128,
+                1000,
                 0,
                 4096,
-                "The minimum distance to spawn the player away from its death location"
+                "Minimum distance for player respawn from the center of the world"
         );
         respawnBlockRange = builder.integerEntry(
                 "respawn_block_range",
-                256,
+                1,
                 0,
                 Integer.MAX_VALUE,
-                "If the player is in this range of its bed/respawn anchor it will respawn there"
+                "Respawn range around bed/respawn anchor"
         );
     }